Clean up WeatherForecaste State to coding standards
using partial classes I want to make it clear what each partial is for.
Base is define what state is there and the constructor.

The behavior is to implement the IState
debug is a for debugging and testing could be conditional compiled
