+++ test/TestApp/TestApp.Client/Features/WeatherForecast/WeatherForecastState.Behavior.cs
@@ -1,0 +1,20 @@
+﻿namespace TestApp.Client.Features.WeatherForecast
+{
+  using BlazorState;
+  using TestApp.Shared.Features.WeatherForecast;
+
+  public partial class WeatherForecastsState : State<WeatherForecastsState>
+  {
+    private WeatherForecastsState(WeatherForecastsState aState) : this()
+    {
+      foreach (WeatherForecastDto weatherForecastDto in aState.WeatherForecasts)
+      {
+        _WeatherForecasts.Add(weatherForecastDto.Clone() as WeatherForecastDto);
+      }
+    }
+
+    public override object Clone() => new WeatherForecastsState(this);
+
+    protected override void Initialize() { }
+  }
+}
